@@ -1,7 +1,7 @@
 import 'package:meta/meta.dart';
 import 'package:flutter/material.dart';
 import 'unit.dart';
-
+const _padding = EdgeInsets.all(16.0);
 class Quotation extends StatefulWidget {
 
 
@@ -19,32 +19,230 @@ class Quotation extends StatefulWidget {
 }
 
 class _QuotationState extends State<Quotation> {
-  
+
+  Unit _fromValue;
+  Unit _toValue;
+  double _inputValue;
+  double _convertedValue = 0;
+  List<DropdownMenuItem> _unitMenuItems;
+  bool _showValidationError = false;
+
   @override
-  Widget build(BuildContext context) {
+  void initState() {
+    super.initState();
+    _createDropdownMenuItems();
+    _setDefaults();
+  }
 
-    final unitWidgets = widget.units.map((Unit unit) {
-      return Container(
-        color: widget.color,
-        margin: EdgeInsets.all(2.0),
-        padding: EdgeInsets.all(12.0),
-        child: Column(
-          children: <Widget>[
-            Text(
-              unit.name,
-              style: Theme.of(context).textTheme.subhead,
-            ),
-            Text(
-              'Price: ${unit.price}'
-            ),
-          ],
+    /// Creates fresh list of [DropdownMenuItem] widgets, given a list of [Unit]s.
+  void _createDropdownMenuItems() {
+    var newItems = <DropdownMenuItem>[];
+    for (var unit in widget.units) {
+      newItems.add(DropdownMenuItem(
+        value: unit.vehicleFactory,
+        child: Container(
+          child: Text(
+            unit.vehicleFactory,
+            softWrap: true,
+          ),
         ),
-      );
-    }).toList();
+      ));
+    }
+    setState(() {
+      _unitMenuItems = newItems;
+    });
+  }
 
+  /// Sets the default values for the 'from' and 'to' [Dropdown]s.
+  void _setDefaults() {
+    setState(() {
+      _fromValue = widget.units[0];
+      _toValue = widget.units[1];
+    });
+  }
 
-    return ListView(
-      children: unitWidgets,
+  /// Clean up conversion; trim trailing zeros, e.g. 5.500 -> 5.5, 10.0 -> 10
+  double _format(double conversion) {
+    var outputNum = conversion.toStringAsPrecision(7);
+    if (outputNum.contains('.') && outputNum.endsWith('0')) {
+      var i = outputNum.length - 1;
+      while (outputNum[i] == '0') {
+        i -= 1;
+      }
+      outputNum = outputNum.substring(0, i + 1);
+    }
+    if (outputNum.endsWith('.')) {
+      return double.parse(outputNum.substring(0, outputNum.length - 1));
+    }
+    return double.parse(outputNum);
+  }
+
+  void _updateConversion() {
+    setState(() {
+      double _fipe = _format(_inputValue);
+      if (_fipe < 30000 ){
+        _convertedValue = 0.00484;
+      } else {
+        _convertedValue = double.parse(((_fipe - 30000) * 0.00181/10000 + 0.00484).toStringAsFixed(5));
+      }
+      print(_toValue.vehicleFactory);
+      if (_fromValue.vehicleFactory == "importado") {
+        _convertedValue = double.parse((_convertedValue * 1.2).toStringAsFixed(5));
+      }
+    });
+  }
+
+  void _updateInputValue(String input) {
+    setState(() {
+      if (input == null || input.isEmpty) {
+        _convertedValue = 0;
+      } else {
+        // Even though we are using the numerical keyboard, we still have to check
+        // for non-numerical input such as '5..0' or '6 -3'
+        try {
+          final inputDouble = double.parse(input);
+          _showValidationError = false;
+          _inputValue = inputDouble;
+          _updateConversion();
+        } on Exception catch (e) {
+          print('Error: $e');
+          _showValidationError = true;
+        }
+      }
+    });
+  }
+
+  Unit _getUnit(String unitName) {
+    return widget.units.firstWhere(
+      (Unit unit) {
+        return unit.vehicleFactory == unitName;
+      },
+      orElse: null,
     );
-  } 
+  }
+
+  void _updateFromConversion(dynamic unitName) {
+    setState(() {
+      _fromValue = _getUnit(unitName);
+    });
+    if (_inputValue != null) {
+      _updateConversion();
+    }
+  }
+
+  void _updateToConversion(dynamic unitName) {
+    setState(() {
+      _toValue = _getUnit(unitName);
+    });
+    if (_inputValue != null) {
+      _updateConversion();
+    }
+  }
+
+    Widget _createDropdown(String currentValue, ValueChanged<dynamic> onChanged) {
+    return Container(
+      margin: EdgeInsets.only(top: 16.0),
+      decoration: BoxDecoration(
+        // This sets the color of the [DropdownButton] itself
+        color: Colors.grey[50],
+        border: Border.all(
+          color: Colors.grey[400],
+          width: 1.0,
+        ),
+      ),
+      padding: EdgeInsets.symmetric(vertical: 8.0),
+      child: Theme(
+        // This sets the color of the [DropdownMenuItem]
+        data: Theme.of(context).copyWith(
+          canvasColor: Colors.grey[50],
+        ),
+        child: DropdownButtonHideUnderline(
+          child: ButtonTheme(
+            alignedDropdown: true,
+            child: DropdownButton(
+              value: currentValue,
+              items: _unitMenuItems,
+              onChanged: onChanged,
+              style: Theme.of(context).textTheme.title,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+  
+    @override
+  Widget build(BuildContext context) {
+    final input = Padding(
+      padding: _padding,
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: [
+          // This is the widget that accepts text input. In this case, it
+          // accepts numbers and calls the onChanged property on update.
+          // You can read more about it here: https://flutter.io/text-input
+          TextField(
+            style: Theme.of(context).textTheme.display1,
+            decoration: InputDecoration(
+              labelStyle: Theme.of(context).textTheme.display1,
+              errorText: _showValidationError ? 'Invalid number entered' : null,
+              labelText: 'FIPE',
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(0.0),
+              ),
+            ),
+            // Since we only want numerical input, we use a number keyboard. There
+            // are also other keyboards for dates, emails, phone numbers, etc.
+            keyboardType: TextInputType.number,
+            onChanged: _updateInputValue,
+          ),
+          _createDropdown(_fromValue.vehicleFactory, _updateFromConversion),
+        ],
+      ),
+    );
+
+    final arrows = RotatedBox(
+      quarterTurns: 1,
+      child: Icon(
+        Icons.compare_arrows,
+        size: 40.0,
+      ),
+    );
+  
+    final minuteValue = Padding(
+      padding: _padding,
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: [
+          InputDecorator(
+            child: Text(
+              '$_convertedValue',
+              style: Theme.of(context).textTheme.display1,
+            ),
+            decoration: InputDecoration(
+              labelText: 'Valor do Minuto',
+              labelStyle: Theme.of(context).textTheme.display1,
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(0.0),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+
+    final converter = Column(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: [
+        input,
+        arrows,
+        minuteValue,
+      ],
+    );
+
+    return Padding(
+      padding: _padding,
+      child: converter,
+    );
+  }
 }
